requirements.txt
asgiref==3.8.1
cellpose==3.0.7
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Django==5.0.4
fastremap==1.14.1
fonttools==4.51.0
idna==3.7
imagecodecs==2024.1.1
imageio==2.34.0
joblib==1.4.0
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.42.0
looseversion==1.3.0
matplotlib==3.8.4
memory-profiler==0.61.0
munkres==1.1.4
natsort==8.4.0
networkx==3.3
numba==0.59.1
numpy==1.24.0
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==1.5.0
patchify==0.2.3
pillow==10.3.0
psutil==5.9.8
pyparsing==3.1.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqtgraph==0.13.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
QSwitchControl==1.0.4
requests==2.31.0
scikit-image==0.23.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sqlparse==0.5.0
threadpoolctl==3.4.0
tifffile==2024.2.12
torch==1.12.0+cu113
torchvision==0.13.0+cu113
tqdm==4.66.2
trackpy==0.6.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
yeastvision==0.1.30